Move TimeLeftDisplay into its own component
